Extract emulator controls into their own module
